+++ apps/sasquatch/src/main/java/com/microsoft/appcenter/sasquatch/activities/MainActivity.java
@@ -112,17 +112,7 @@ public class MainActivity extends AppCompatActivity {
         /* Set the track explicitly only if we set it in settings, to test the initial public by default at first launch. */
         int savedTrack = sSharedPreferences.getInt(application.getString(R.string.appcenter_distribute_track_state_key), 0);
         if (savedTrack != 0) {
-            try {
-
-                /*
-                 * TODO replace the next line with 'Distribute.setUpdateTrack(savedTrack);'
-                 * when updating the demo during release process.
-                 */
-                Method setUpdateTrackMethod = Distribute.class.getMethod("setUpdateTrack", int.class);
-                setUpdateTrackMethod.invoke(null, savedTrack);
-            } catch (Exception e) {
-                Toast.makeText(application, "No Update Track API in this build", Toast.LENGTH_SHORT).show();
-            }
+            Distribute.setUpdateTrack(savedTrack);
         }
         boolean automaticCheckForUpdate = sSharedPreferences.getBoolean(application.getString(R.string.appcenter_distribute_disable_check_for_update_key), true);
         if (!automaticCheckForUpdate) {