+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # App Center SDK for Android Change Log
 
-## Version 3.0.0 (Under development)
+## Version 3.1.0 (Under development)
+
+### App Center Distribute
+
+* **[Feature]** Add a `disableAutomaticCheckForUpdate` API that needs to be called before SDK start in order to turn off automatic check for update. 
+* **[Feature]** Add a `checkForUpdate` API to manually check for update.
+
+## Version 3.0.0
 
 ### App Center Auth